Add feedback trait
Define a new generic trait for any Weave client to request feedback from
a device. Feedback in that instance corresponds to on-device diagnostic
data used to debug/investigate issues (stack trace, memory dump, sensor
data captures, log files, etc).

This trait does not impose any restrictions on the data format, the kind
of data that can be retrieved nor the protocols by which the data is
uploaded.

The act of requesting feedback does not alter any logging levels on the
device (at least not by requirement).